ADD: conexion completa CRUD, verificacion de logeo, verificacion de sesion MOD: seeder de usuario

+++ Laravel/app/Http/Controllers/ControladorLibro.php
@@ -1,37 +1,61 @@
 <?php
 
 namespace App\Http\Controllers;
+
 use Illuminate\Http\Request;
 use App\Models\Libro;
-
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 class ControladorLibro extends Controller
 {
     public function index(){
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         $libros = Libro::all();
-        return view('proyecto.login')->with('libros', $libros);
+        return view('proyecto.inicio')->with('libros', $libros);
     }
-    public function crear(){
-        return view('proyecto.añadir');
+    public function create(){
+        if (!Auth::check()) {
+            return redirect('login');
+        }
+        return view('proyecto.agregar');
     }
-    public function guardar(Request $request){
+    public function store(Request $request){
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         $libro= new Libro;
-        $libro->Nombre= $request->input('nombre');
+        $libro->Nombre= $request->input('nombrelibro');
         $libro->Autor= $request->input('autor');
         $libro->ISBN= $request->input('isbn');
+        $libro->Imagen= $request->input('img');
         $libro->save();
         return redirect('inicio');
     }
-    public function editar($id){
+    public function edit($id){
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         $libro = Libro::find($id);
         return view('proyecto.modificar')->with('libro',$libro);
     }
-    public function modificar (Request $request, $id){
-        $libro->Nombre= $request->input('nombre');
+    public function update(Request $request, $id){
+        if (!Auth::check()) {
+            return redirect('login');
+        }
+        $libro = Libro::find($id);
+        $libro->Nombre= $request->input('nombrelibro');
         $libro->Autor= $request->input('autor');
         $libro->ISBN= $request->input('isbn');
+        $libro->Imagen= $request->input('img');
         $libro->save();
+        return redirect('inicio');
     }
-    public function eliminar($id){
+    public function destroy($id){
+        if (!Auth::check()) {
+            return redirect('login');
+        }
         Libro::destroy($id);
         return redirect('inicio');
     }
